chore(proto): rename proto path back to sf
