+++ cmd/firearweave/cli/mindreader-node.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/ChainSafe/firehose-arweave/codec"
-	pbcodec "github.com/ChainSafe/firehose-arweave/pb/cs/arweave/codec/v1"
+	pbcodec "github.com/ChainSafe/firehose-arweave/pb/sf/arweave/type/v1"
 	"github.com/spf13/cobra"
 	"github.com/streamingfast/bstream"
 	"github.com/streamingfast/bstream/blockstream"
