+++ codec/decoder.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 	"google.golang.org/protobuf/proto"
 
-	pbcodec "github.com/ChainSafe/firehose-arweave/pb/cs/arweave/codec/v1"
+	pbcodec "github.com/ChainSafe/firehose-arweave/pb/sf/arweave/type/v1"
 	"github.com/streamingfast/bstream"
 	pbbstream "github.com/streamingfast/pbgo/sf/bstream/v1"
 )
