+++ tools/blocks.go
@@ -21,7 +21,7 @@ import (
 	"io"
 	"strconv"
 
-	pbcodec "github.com/ChainSafe/firehose-arweave/pb/cs/arweave/codec/v1"
+	pbcodec "github.com/ChainSafe/firehose-arweave/pb/sf/arweave/type/v1"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/streamingfast/bstream"
